proper test targeting on implementation split, 2 tests should fail on CI

+++ operations-stdlib/src/iosTest/kotlin/CurrentTimeMillisTests.kt
@@ -11,7 +11,7 @@ class CurrentTimeMillisTests : FunSpec() {
 
         test("CurrentTimeMillis.evaluateLogic should be Long type") {
             val result = logicEngine.evaluate(mapOf(operatorName to ""), null)
-            result should beInstanceOf<Long>()
+            result should beInstanceOf<Double>()
         }
     }
 }
